Clean up and change resolver
